JAMES-2544 Use registered extensions for Cassandra event sourcing systems
